agents/calc-agent/requirements.txt
flask>=3.0.0
fastapi>=0.100.0
starlette>=0.27.0
uvicorn[standard]>=0.27.0
a2wsgi>=1.7.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.27.0
fastmcp>=2.0
gt8004-sdk[mcp] @ git+https://github.com/HydroX-labs/gt8004-sdk.git
